Add explanation of certificate re-import, refactor namespace to prevent name conflicts
