+++ MMS/Services/Matchmaking/JoinSessionMessenger.cs
@@ -1,5 +1,5 @@
 using System.Net.WebSockets;
-using MMS.Models.Lobby;
+using MMS.Models.Lobbies;
 using MMS.Models.Matchmaking;
 using MMS.Services.Lobbies;
 using MMS.Services.Network;
